+++ crypto/fips1402/src/main/java/org/keycloak/crypto/fips/BCFIPSCertificateUtilsProvider.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.keycloak.crypto.def;
+package org.keycloak.crypto.fips;
 
 import org.bouncycastle.asn1.x500.X500Name;
 import org.bouncycastle.asn1.x509.BasicConstraints;
@@ -32,8 +32,7 @@ import org.bouncycastle.cert.jcajce.JcaX509ExtensionUtils;
 import org.bouncycastle.operator.ContentSigner;
 import org.bouncycastle.operator.jcajce.JcaContentSignerBuilder;
 import org.keycloak.common.util.BouncyIntegration;
-import org.keycloak.common.crypto.CertificateUtils;
-
+import org.keycloak.common.crypto.CertificateUtilsProvider;
 
 import java.math.BigInteger;
 import java.security.KeyPair;
@@ -50,7 +49,7 @@ import java.util.Date;
  * @author <a href="mailto:giriraj.sharma27@gmail.com">Giriraj Sharma</a>
  * @version $Revision: 2 $
  */
-public class CertificateUtilsBC implements CertificateUtils {
+public class BCFIPSCertificateUtilsProvider implements CertificateUtilsProvider{
 
     /**
      * Generates version 3 {@link java.security.cert.X509Certificate}.
